# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go语言和fiber框架的高性能高并发的Web项目骨架",
        "title": "Go-skeleton API Docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9528",
    "basePath": "/",
    "paths": {
...
        "/farm/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场日志"
                ],
                "summary": "运行日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "accountId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模块 farm|friend|system",
                        "name": "module",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "条数上限",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场日志"
                ],
                "summary": "清空运行日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID（空则清空全部）",
                        "name": "accountId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/status/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场状态"
                ],
                "summary": "账号运行状态详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账号ID",
                        "name": "accountId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/farm/status/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "农场状态"
                ],
                "summary": "账号运行状态列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称/编码关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/system/admin/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "创建租户用户",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.CreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/admin/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账号/昵称关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 1正常 2禁用",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/admin/modify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新租户用户",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/admin/status": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "启停租户用户",
                "parameters": [
                    {
                        "description": "启停参数",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.StatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.CreateReq": {
            "type": "object",
            "required": [
                "account",
                "nickName",
                "password",
                "status"
            ],
            "properties": {
                "account": {
                    "type": "string",
                    "maxLength": 64
                },
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "realName": {
                    "type": "string"
                },
                "roleIds": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.StatusReq": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "github_com_it00021hot_qq-farm-core_internal_app_controller_system_admin.UpdateReq": {
            "type": "object",
            "required": [
                "id",
                "nickName",
                "status"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickName": {
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "realName": {
                    "type": "string"
                },
                "roleIds": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "response.Code": {
            "type": "integer",
            "enum": [
                0,
                10001,
                10002,
                10003,
                10004,
                10005,
                10006,
                10007,
                10008
            ],
            "x-enum-varnames": [
                "Success",
                "Failed",
                "UnAuthed",
                "AuthExpired",
                "InternalErr",
                "RequestMethodErr",
                "RequestParamErr",
                "RequestFormErr",
                "AccountInfoErr"
            ]
        },
        "response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.Code"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}